:truck: [RENAME] 폴더구조 개선 #39
