+++ src/main/java/com/example/eight/global/oauth2/controller/OauthController.java
@@ -3,8 +3,8 @@ package com.example.eight.global.oauth2.controller;
 import com.example.eight.global.jwt.JwtService;
 import com.example.eight.global.oauth2.service.OauthService;
 import com.example.eight.user.SocialType;
-import com.example.eight.user.dto.JwtTokenDto;
-import com.example.eight.user.dto.ResponseDto;
+import com.example.eight.global.jwt.dto.JwtTokenDto;
+import com.example.eight.global.ResponseDto;
 import com.example.eight.global.oauth2.dto.UserInfoDto;
 import com.example.eight.user.service.UserService;
 import lombok.RequiredArgsConstructor;
